% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{raster2mat}
\alias{raster2mat}
\title{Write a Raster to a mat file}
\usage{
raster2mat(x, filename = "raster2mat.mat", overwrite = TRUE, compress = 0)
}
\arguments{
\item{x}{the Raster* object}

\item{filename}{the fully qualified filename to write to}

\item{overwrite}{logical, if FALSE then stop if filename already exists}

\item{compress}{numeric if non-zero then gzip the result at the compression
level specified}
}
\value{
the original Raster* invisibly
}
\description{
The output file captures basic elements of the Raster
 \itemize{
     \item{data the variable data}
     \item{projection charcate string}
     \item{locations list of cols, rows and z}
 }
}

