% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_matlab}
\alias{read_matlab}
\title{Read a matlab file and return the first element or the specified element}
\usage{
read_matlab(filename, elem = 1)
}
\arguments{
\item{filename}{the fully qualified file path}

\item{elem}{character or numeric index of the element to return}
}
\value{
the item at index 'elem' unless there is an error in which case NULL
}
\description{
Read a matlab file and return the first element or the specified element
}

