% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_matlab}
\alias{write_matlab}
\title{Write a matlab file}
\usage{
write_matlab(x, filename)
}
\arguments{
\item{x}{the object to write}

\item{filename}{the fully qualified path to the file}
}
\value{
value returned by \code{R.matlab::writeMat()}
}
\description{
Write a matlab file
}

